+++ CHANGELOG.md
@@ -60,6 +60,7 @@ In this release we had a little focus on:
 - Forbids to use ``type()`` for compares
 - Forbids to have consecutive expressions with too deep access level
 - Forbids to have too many public instance attributes
+- Forbids to use pointless star operations: `print(*[])`
 
 ### Bugfixes